requirements.txt
Django==5.0.6
djangorestframework==3.15
psycopg2-binary==2.9.9
django-filter==24.2
psycopg2>=2.9,<3
python-dotenv
drf-yasg

# Dev requirements
pytest
pytest-cov
pytest-mock
pytest-django
